Restructure the two implementation
