+++ crates/calibration/src/center_circle/measurement.rs
@@ -1,7 +1,8 @@
-use super::circles::CenterOfCircleAndPoints;
 use coordinate_systems::Pixel;
 use projection::camera_matrix::CameraMatrix;
 use types::camera_position::CameraPosition;
+
+use super::circles::CenterOfCircleAndPoints;
 
 #[derive(Clone)]
 pub struct Measurement {
