+++ crates/calibration/src/goal_box/lines.rs
