+++ crates/calibration/src/goal_box/mod.rs
@@ -1,0 +1,3 @@
+pub mod lines;
+pub mod measurement;
+pub mod residuals;
